+++ src/components/display/GameServer/EditGameServer/McRouterDomainSelectorEdit.tsx
@@ -23,7 +23,7 @@ const McRouterDomainSelectorEdit = ({
   externalGameId,
 }: McRouterDomainSelectorEditProps) => {
   const { t } = useTranslationPrefix("mcRouterDomains");
-  const { uuid, role } = useContext(AuthContext);
+  const { uuid } = useContext(AuthContext);
   const [isOpen, setIsOpen] = useState(false);
 
   const mcRouterConfig = useTypedSelector(
@@ -46,17 +46,13 @@ const McRouterDomainSelectorEdit = ({
 
     const allDomains = mcRouterConfig.domains;
 
-    if (role === "ADMIN" || role === "OWNER") {
-      return allDomains;
-    }
-
     if (currentUser?.mc_router_allow_all_domains) {
       return allDomains;
     }
 
     const allowedDomains = currentUser?.mc_router_allowed_domains ?? [];
     return allDomains.filter((domain) => allowedDomains.includes(domain));
-  }, [isMcRouterEnabled, mcRouterConfig?.domains, role, currentUser]);
+  }, [isMcRouterEnabled, mcRouterConfig?.domains, currentUser]);
 
   const handleDomainToggle = useCallback(
     (domain: string) => {
